requirements.txt
alembic==1.17.2
blinker==1.9.0
brotli==1.2.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cssselect2==0.8.0
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.61.1
greenlet==3.3.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
openpyxl==3.1.5
pillow==12.1.0
pycparser==2.23
pydyf==0.12.1
pyphen==0.17.2
SQLAlchemy==2.0.45
tinycss2==1.5.1
tinyhtml5==2.0.0
typing_extensions==4.15.0
weasyprint==67.0
webencodings==0.5.1
Werkzeug==3.1.4
WTForms==3.2.1
zopfli==0.4.0
